( +interface )

+++ src/ru/fizteh/fivt/students/valentinbarishev/filemap/ShellDbRemove.java
@@ -3,9 +3,9 @@ package ru.fizteh.fivt.students.valentinbarishev.filemap;
 import ru.fizteh.fivt.students.valentinbarishev.shell.SimpleShellCommand;
 
 public final class ShellDbRemove extends SimpleShellCommand {
-    private DataBase dataBase;
+    private SimpleDataBase dataBase;
 
-    public ShellDbRemove(final DataBase newDataBase) {
+    public ShellDbRemove(final SimpleDataBase newDataBase) {
         setName("remove");
         setNumberOfArgs(2);
         setHint("usage: remove <key>");
@@ -14,6 +14,10 @@ public final class ShellDbRemove extends SimpleShellCommand {
 
     @Override
     public void run() {
+        if (!dataBase.exist()) {
+            System.out.println("no table");
+            return;
+        }
         if (!dataBase.remove(getArg(1))) {
             System.out.println("not found");
         } else {
